pyproject.toml
[project]
name = "office"
version = "0.1.0"
description = "办公小工具整合平台 - AI对话操作 + 节点流编辑 + 插件化扩展"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    # UI框架
    "PySide6>=6.6.0",
    # AI Agent框架
    "agentscope>=1.0.17",
    # 持久化层
    "sqlalchemy>=2.0",
    "aiosqlite>=0.19.0",
    # 加密工具 (API Key加密)
    "cryptography>=41.0",
    # Git操作 (节点包管理)
    "GitPython>=3.1",
    # 文档处理
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "pymupdf>=1.26.3",
    # 配置管理
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-qt>=4.2.0",
    "ruff>=0.1.0",
]

[project.scripts]
office = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]

[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-qt>=4.5.0",
]
